Rename error-handling.js to print-pretty-error.js, and create test
